requirements.txt
langgraph>=0.3.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
mcp>=1.0.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-dotenv>=1.0.0
rich>=13.0.0
